pyproject.toml
[tool.poetry]
name = "social-insecurity"
version = "2.0.0"
description = "A website written in Flask for practicing patching security flaws and bugs"
authors = ["solbero <njosol@pm.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = {extras = ["dotenv"], version = "^3.0.0"}
Flask-WTF = "^1.2.0"
pytest = "^8.0.0"
bcrypt = "^4.2.0"
flask-login = "^0.6.3"
flask-bcrypt = "^1.0.1"
flask-sqlalchemy = "^3.1.1"

[tool.poetry.group.dev.dependencies]
djlint = "^1.34.0"
tox = "^4.0.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.tox]
legacy_tox_ini = """
    [tox]
    envlist = py39, py310, py311, py312

    [testenv]
    skip_install = true
    allowlist_externals = poetry
    commands_pre = poetry install
    commands = pytest {posargs}
"""

[tool.ruff]
line-length = 120
target-version = 'py39'
lint.select=["I"]

[tool.djlint]
indent = 2
max_line_length=120
format_css=true
format_js=true
profile="jinja"
extension="html.j2"
